requirements.txt
alembic==1.4.3
blinker==1.4
certifi==2020.6.20
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.0
Mako==1.1.3
numpy==1.17.4
psycopg2-binary==2.8.6
pymongo==3.11.0
pytesseract==0.3.2
python-dateutil==2.8.1
python-editor==1.0.4
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
